+++ specs/config.prettier.specs.js
@@ -1,10 +1,18 @@
+/**
+ * Copyright (c) 2019, Patricio Trevino
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 "use strict";
 
-const childProcess = require("child_process");
+const spawn = require("cross-spawn");
 
 describe("config", () => {
-  it("samples does not causes errors without eslint-config-typescript/prettier", () => {
-    const result = childProcess.spawnSync(
+  it("samples does not throw errors without eslint-config-typescript/prettier", () => {
+    const result = spawn.sync(
       "npm",
       [
         "run",
@@ -27,8 +35,8 @@ describe("config", () => {
     expect(output[0].messages.length).toBe(0);
   });
 
-  it("samples causes errors with eslint-config-typescript/prettier", () => {
-    const result = childProcess.spawnSync(
+  it("samples throws errors with eslint-config-typescript/prettier", () => {
+    const result = spawn.sync(
       "npm",
       [
         "run",
@@ -48,21 +56,18 @@ describe("config", () => {
     );
 
     expect(output.length).toBe(1);
-    expect(output[0].messages.length).toBe(1);
+    expect(output[0].messages.length).toBe(2);
+    expect(output[0].messages[0].message).toBe(
+      "Replace `'one'` with `\"one\"`"
+    );
+    expect(output[0].messages[1].message).toBe("Insert `;`");
   });
 
-  it("returns all rules", () => {
+  it("returns configuration", () => {
     const config = require("../prettier");
 
     expect(config).toBeDefined();
-    expect(config.extends).toContain("prettier");
-    expect(config.extends).toContain("prettier/react");
-    expect(config.plugins).toContain("prettier");
-    expect(config.rules).toBeDefined();
-    expect(config.rules["prettier/prettier"]).toBeDefined();
-    expect(config.rules["prettier/prettier"][0]).toBe("error");
-    expect(config.rules["prettier/prettier"][1]).toMatchObject({
-      parser: "typescript"
-    });
+    expect(config.extends).toContain("plugin:prettier/recommended");
+    expect(config.extends).toContain("prettier/typescript");
   });
 });