+++ specs/config.react.specs.js
@@ -1,12 +1,18 @@
+/**
+ * Copyright (c) 2019, Patricio Trevino
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 "use strict";
 
-const fs = require("fs");
-const path = require("path");
-const childProcess = require("child_process");
+const spawn = require("cross-spawn");
 
 describe("config", () => {
-  it("samples does not causes errors without eslint-config-typescript/react", () => {
-    const result = childProcess.spawnSync(
+  it("samples does not throw errors without eslint-config-typescript/react", () => {
+    const result = spawn.sync(
       "npm",
       [
         "run",
@@ -29,8 +35,8 @@ describe("config", () => {
     expect(output[0].messages.length).toBe(0);
   });
 
-  it("samples causes errors with eslint-config-typescript/react", () => {
-    const result = childProcess.spawnSync(
+  it("samples throws errors with eslint-config-typescript/react", () => {
+    const result = spawn.sync(
       "npm",
       [
         "run",
@@ -51,13 +57,15 @@ describe("config", () => {
 
     expect(output.length).toBe(1);
     expect(output[0].messages.length).toBe(3);
+    expect(output[0].messages[0].ruleId).toBe("react/display-name");
+    expect(output[0].messages[1].ruleId).toBe("react/no-is-mounted");
+    expect(output[0].messages[2].ruleId).toBe("react/react-in-jsx-scope");
   });
 
-  it("returns all rules", () => {
+  it("returns configuration", () => {
     const config = require("../react");
 
     expect(config).toBeDefined();
     expect(config.extends).toContain("plugin:react/recommended");
-    expect(config.plugins).toContain("react");
   });
 });